grouped tests into modules
