+++ tests/test_hyperopt/test_hyperopt.py
