+++ tests/test_models/test_optimize.py
@@ -1,209 +1,27 @@
+
 import os
-import time
 import unittest
 
-import numpy as np
-import tensorflow as tf
-
-if 230 <= int(''.join(tf.__version__.split('.')[0:2]).ljust(3, '0')) < 250:
-    from ai4water.functional import Model
-    print(f"Switching to functional API due to tensorflow version {tf.__version__}")
-else:
-    from ai4water import Model
-
+from ai4water import Model
+from ai4water.preprocessing import DataHandler
 from ai4water.datasets import arg_beach
-from ai4water.utils.utils import find_best_weight
-from ai4water.functional import Model as FModel
-from ai4water.preprocessing.datahandler import DataHandler
 from ai4water._optimize import make_space
 from ai4water.utils.utils import process_config_dict
-from ai4water.utils.utils import find_opt_paras_from_model_config, prepare_data
+from ai4water.utils.utils import find_opt_paras_from_model_config
 from ai4water.hyperopt import Categorical, Real, Integer
 
-data = arg_beach()
-dh = DataHandler(data=data, verbosity=0)
-x, y = dh.training_data()
 
-mlp_model = {"layers": {"Dense": 8, "Dense_1": 1}}
+data = arg_beach()
 
 
 def get_lstm():
 
-    m_conf = {"layers": {"LSTM": {"config": {"units": Integer(2, 5, num_samples=10)}},
-                         "relu": {},
-                         "Dense_0": {"units": Integer(2, 5, name="dense1_units", num_samples=10),
-                                     "activation": Categorical(["relu", "tanh"], name="dense1_act")},
-                         "Dense_1": {"config": {"units": 1, "activation": "relu"}}}}
-    return m_conf
+    return {"layers": {"LSTM": {"config": {"units": Integer(2, 5, num_samples=10)}},
+                       "relu": {},
+                       "Dense_0": {"units": Integer(2, 5, name="dense1_units", num_samples=10),
+                                   "activation": Categorical(["relu", "tanh"], name="dense1_act")},
+                       "Dense_1": {"config": {"units": 1, "activation": "relu"}}}}
 
-
-def test_user_defined_data(_model):
-    # using user defined x
-    t, p = _model.predict(x=x, return_true=True)
-    assert t is None
-    assert len(p) == len(x)
-
-    # using user defined x and y, post_processing must happen
-    t, p = _model.predict(x=x, y=y, return_true=True)
-    assert len(t) == len(p) == len(y)
-
-    return
-
-
-def _test_ml_inbuilt_data(_model):
-    model = _model(model="RandomForestRegressor",
-                   data=arg_beach(),
-                   verbosity=0)
-    model.fit()
-
-    test_user_defined_data(model)
-
-    t, p = model.predict(return_true=True)
-    assert len(t) == len(p)
-    return
-
-
-def _test_ml_userdefined_data(_model):
-    model = _model(model="RandomForestRegressor", verbosity=0)
-    model.fit(x=x, y=y)
-
-    test_user_defined_data(model)
-
-    return model
-
-
-def _test_fit(_model):
-    model = _model(model="RandomForestRegressor", verbosity=0)
-    model.fit(x, y)
-
-    model.fit(x, y=y)
-
-    model.fit(x=x, y=y)
-    return
-
-
-def _test_ml_userdefined_non_kw(_model):
-    # using non-keyword arguments to .predict
-    model = _model(model="RandomForestRegressor", verbosity=0)
-    model.fit(x=x, y=y)
-
-    model.predict(x)
-    return
-
-
-def _test_hydro_metrics(_model):
-    model = _model(model="RandomForestRegressor", verbosity=0)
-    model.fit(x=x, y=y)
-
-    for metrics in ["minimal", "hydro_metrics", "all"]:
-        model.predict(x=x, metrics=metrics)
-    return
-
-
-def _test_from_config_basic(_model, find_best=False, config_file=False):
-
-    for m in ["RandomForestRegressor",
-              "XGBRegressor",
-              "CatBoostRegressor",
-              "LGBMRegressor",
-              mlp_model
-              ]:
-        model = _model(model=m, data=arg_beach(), lookback=1, verbosity=0)
-        model.fit(x, y)
-        ini_y = model.predict(np.arange(13).reshape(-1, 13)).item()
-
-        if config_file:
-            m2 = Model.from_config_file(os.path.join(model.path, 'config.json'))
-        else:
-            m2 = Model.from_config(model.config)
-
-        best_weight = None
-        if find_best:
-            best_weight = os.path.join(model.w_path, find_best_weight(model.w_path))
-
-        m2.update_weights(best_weight)
-        fin_y = m2.predict(np.arange(13).reshape(-1, 13)).item()
-        assert np.allclose(ini_y, fin_y)
-        time.sleep(1)
-
-    return
-
-
-class TestPredictMethod(unittest.TestCase):
-
-    def test_ml_inbuilt_data(self):
-        _test_ml_inbuilt_data(Model)
-        return
-
-    def test_ml_inbuilt_data_fn(self):
-        _test_ml_inbuilt_data(FModel)
-        return
-
-    def test_ml_userdefined_data(self):
-        model = _test_ml_userdefined_data(Model)
-        # using data generated by DataHnadler
-        self.assertRaises(ValueError, model.predict)
-        return
-
-    def test_ml_userdefined_data_fn(self):
-        model = _test_ml_userdefined_data(FModel)
-        # using data generated by DataHnadler
-        self.assertRaises(ValueError, model.predict)
-        return
-
-    def test_ml_userdefined_non_kw(self):
-        _test_ml_userdefined_non_kw(Model)
-        return
-
-    def test_ml_userdefined_non_kw_fn(self):
-        _test_ml_userdefined_non_kw(FModel)
-        return
-
-    def test_hydro_metrics(self):
-        _test_hydro_metrics(Model)
-        return
-
-    def test_hydro_metrics_functional(self):
-        _test_hydro_metrics(FModel)
-        return
-
-
-class TestFit(unittest.TestCase):
-
-    def test_fit(self):
-        _test_fit(Model)
-        return
-
-    def test_fit_functional(self):
-        _test_fit(FModel)
-        return
-
-
-class TestFromConfig(unittest.TestCase):
-
-    def test_subclassing(self):
-        _test_from_config_basic(Model)
-        return
-
-    def test_subclassing_fn(self):
-        _test_from_config_basic(FModel)
-
-    def test_subclassing_with_weights(self):
-        _test_from_config_basic(Model, find_best=True)
-
-    def test_subclassing_fn_with_weights(self):
-        # we are able to load functinoal model
-        _test_from_config_basic(FModel, find_best=True)
-
-    def test_subclassing_with_config_file(self):
-        # we are able to load subclassing Model from config_file
-        _test_from_config_basic(Model, config_file=True)
-        return
-
-    def test_fn_with_config_file(self):
-        # we are able to load functional model from config_file
-        _test_from_config_basic(FModel, config_file=True)
-        return
 
 
 class TestOptimize(unittest.TestCase):
@@ -528,72 +346,6 @@ class TestOptimizeHyperparas(unittest.TestCase):
         optimizer = model.optimize_hyperparameters(algorithm="random", num_iterations=5, process_results=False)
         assert model.config['model']['layers']['LSTM']['config']['units'] == optimizer.best_paras()['units']
         assert model.config['lookback'] == optimizer.best_paras()['lookback']
-        return
-
-
-class TestEvaluate(unittest.TestCase):
-    model = Model(
-        model={"layers": {"Dense": 1}},
-        lookback=1,
-        data=arg_beach(),
-        verbosity=0,
-    )
-
-    def test_basic(self):
-        eval_scores = self.model.evaluate()
-        assert isinstance(eval_scores, list) and len(eval_scores) == 2
-        return
-
-    def test_basic_with_metrics(self):
-        # basic example with metrics
-        eval_scores = self.model.evaluate(metrics="kge")
-        assert isinstance(eval_scores, float)
-
-    def test_basic_with_metric_groups(self):
-        # basic example with metrics
-        eval_scores = self.model.evaluate(metrics="hydro_metrics")
-        assert isinstance(eval_scores, dict)
-        return
-
-    def test_custom_xy(self):
-        eval_scores = self.model.evaluate(np.random.random((10, 13)), np.random.random((10, 1, 1)))
-        assert isinstance(eval_scores, list) and len(eval_scores) == 2
-        return
-
-    def test_custom_xy0(self):
-        # only y as keyword
-        eval_scores = self.model.evaluate(np.random.random((10, 13)), y=np.random.random((10, 1, 1)))
-        assert isinstance(eval_scores, list) and len(eval_scores) == 2
-        return
-
-    def test_custom_xy_as_keyword_args(self):
-        eval_scores = self.model.evaluate(x=np.random.random((10, 13)), y=np.random.random((10, 1, 1)))
-        assert isinstance(eval_scores, list) and len(eval_scores) == 2
-        return
-
-    def test_custom_xy_with_metrics(self):
-        # custom data with metrics
-        pbias = self.model.evaluate(x=np.random.random((10, 13)), y=np.random.random((10, 1, 1)),
-                                    metrics='pbias')
-        assert isinstance(pbias, float)
-        return
-
-    def test_custom_xy_with_metric_groups(self):
-        # custom data with group of metrics
-        hydro = self.model.evaluate(x=np.random.random((10, 13)), y=np.random.random((10, 1, 1)),
-                                    metrics='hydro_metrics')
-        assert isinstance(hydro, dict)
-        return
-
-class TestPermImp(unittest.TestCase):
-
-    def test_basic0(self):
-        model = Model(model="XGBRegressor",
-                      verbosity=0,
-                      data=arg_beach())
-        model.fit()
-        imp = model.permutation_importance(data="validation")
-        assert imp.shape == (13, 5)
         return
