+++ tests/test_models/test_pytorch/test_declarative_nn_torch.py
@@ -68,6 +68,3 @@ class TestTorchDeclarativeDef(unittest.TestCase):
 
 if __name__ == "__main__":
     unittest.main()
-
-
-
